requirements.txt
tortoise-orm==0.19.0
asyncpg==0.25.0